+++ packages/monaco-editor-wrapper/src/editorAppExtended.ts
@@ -1,9 +1,12 @@
 import type * as vscode from 'vscode';
-import { EditorAppBase, EditorAppBaseConfig, EditorAppType } from './editorAppBase.js';
+import { IDisposable, editor } from 'monaco-editor';
+import getThemeServiceOverride from '@codingame/monaco-vscode-theme-service-override';
+import getTextmateServiceOverride from '@codingame/monaco-vscode-textmate-service-override';
+import { whenReady as whenReadyTheme } from '@codingame/monaco-vscode-theme-defaults-default-extension';
+import { EditorAppBase, EditorAppConfigBase, ModelUpdateType, isModelUpdateRequired } from './editorAppBase.js';
 import { registerExtension, IExtensionManifest, ExtensionHostKind } from 'vscode/extensions';
 import { UserConfig } from './wrapper.js';
 import { verifyUrlorCreateDataUrl } from './utils.js';
-import { IDisposable } from 'monaco-editor';
 import { Logger } from './logger.js';
 
 export type ExtensionConfig = {
@@ -11,9 +14,14 @@ export type ExtensionConfig = {
     filesOrContents?: Map<string, string | URL>;
 };
 
-export type EditorAppConfigVscodeApi = EditorAppBaseConfig & {
-    $type: 'vscodeApi';
+export type UserConfiguration = {
+    json?: string;
+}
+
+export type EditorAppConfigExtended = EditorAppConfigBase & {
+    $type: 'extended';
     extensions?: ExtensionConfig[];
+    userConfiguration?: UserConfiguration;
 };
 
 export type RegisterExtensionResult = {
@@ -34,25 +42,22 @@ export type RegisterLocalProcessExtensionResult = RegisterLocalExtensionResult &
 /**
  * The vscode-apo monaco-editor app uses vscode user and extension configuration for monaco-editor.
  */
-export class EditorAppVscodeApi extends EditorAppBase {
+export class EditorAppExtended extends EditorAppBase {
 
-    private config: EditorAppConfigVscodeApi;
+    private config: EditorAppConfigExtended;
     private extensionRegisterResults: Map<string, RegisterLocalProcessExtensionResult | RegisterExtensionResult | undefined> = new Map();
     private logger: Logger | undefined;
 
     constructor(id: string, userConfig: UserConfig, logger?: Logger) {
         super(id);
         this.logger = logger;
-        this.config = this.buildConfig(userConfig) as EditorAppConfigVscodeApi;
-        const userInput = userConfig.wrapperConfig.editorAppConfig as EditorAppConfigVscodeApi;
-        this.config.extensions = userInput.extensions ?? undefined;
+        const userAppConfig = userConfig.wrapperConfig.editorAppConfig as EditorAppConfigExtended;
+        this.config = this.buildConfig(userAppConfig) as EditorAppConfigExtended;
+        this.config.extensions = userAppConfig.extensions ?? undefined;
+        this.config.userConfiguration = userAppConfig.userConfiguration ?? undefined;
     }
 
-    getAppType(): EditorAppType {
-        return 'vscodeApi';
-    }
-
-    getConfig(): EditorAppConfigVscodeApi {
+    getConfig(): EditorAppConfigExtended {
         return this.config;
     }
 
@@ -60,17 +65,18 @@ export class EditorAppVscodeApi extends EditorAppBase {
         return this.extensionRegisterResults.get(extensionName);
     }
 
-    async createEditors(container: HTMLElement): Promise<void> {
-        if (this.config.useDiffEditor) {
-            await this.createDiffEditor(container);
-        } else {
-            await this.createEditor(container);
-        }
+    override specifyServices(): editor.IEditorOverrideServices {
+        return {
+            ...getThemeServiceOverride(),
+            ...getTextmateServiceOverride()
+        };
     }
 
-    async init() {
-        // await all extenson that should be ready beforehand
-        await this.awaitReadiness(this.config.awaitExtensionReadiness);
+    override async init() {
+        // await all extensions that should be ready beforehand
+        // always await theme extension
+        const awaitReadiness = (this.config.awaitExtensionReadiness ?? []).concat(whenReadyTheme);
+        await this.awaitReadiness(awaitReadiness);
 
         if (this.config.extensions) {
             const allPromises: Array<Promise<void>> = [];
@@ -89,13 +95,27 @@ export class EditorAppVscodeApi extends EditorAppBase {
         }
 
         // buildConfig ensures userConfiguration is available
-        await this.updateUserConfiguration(this.config.userConfiguration);
-        this.logger?.info('Init of VscodeApiConfig was completed.');
+        await this.updateUserConfiguration(this.config.userConfiguration?.json);
+        this.logger?.info('Init of Extended App was completed.');
     }
 
     disposeApp(): void {
         this.disposeEditor();
         this.disposeDiffEditor();
         this.extensionRegisterResults.forEach((k) => k?.dispose());
+    }
+
+    isAppConfigDifferent(orgConfig: EditorAppConfigExtended, config: EditorAppConfigExtended, includeModelData: boolean): boolean {
+        let different = false;
+        if (includeModelData) {
+            different = isModelUpdateRequired(orgConfig, config) !== ModelUpdateType.NONE;
+        }
+        const propsExtended = ['useDiffEditor', 'domReadOnly', 'readOnly', 'awaitExtensionReadiness', 'overrideAutomaticLayout', 'editorOptions', 'diffEditorOptions', 'userConfiguration', 'extensions'];
+        type ExtendedKeys = keyof typeof orgConfig;
+        const propCompareExtended = (name: string) => {
+            return orgConfig[name as ExtendedKeys] !== config[name as ExtendedKeys];
+        };
+        different = different || propsExtended.some(propCompareExtended);
+        return different;
     }
 }